Extract all fixtures to their own namespace.
