+++ tests/Fixtures/BananaException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Laravel\Telescope\Tests\Fixtures;
+
+use Exception;
+
+class BananaException extends Exception
+{
+    //
+}
