+++ tests/Fixtures/DummyEvent.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Laravel\Telescope\Tests\Fixtures;
+
+class DummyEvent
+{
+    public $data;
+
+    public function __construct(...$payload)
+    {
+        $this->data = $payload;
+    }
+
+    public function handle()
+    {
+        //
+    }
+}
