+++ tests/Fixtures/MyCommand.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Laravel\Telescope\Tests\Fixtures;
+
+use Illuminate\Console\Command;
+
+class MyCommand extends Command
+{
+    protected $signature = 'telescope:test-command';
+
+    public function handle()
+    {
+        //
+    }
+}
